+++ packages/perseus-editor/src/widgets/image-editor/components/image-dimensions-input.tsx
@@ -1,0 +1,88 @@
+import {LabeledField} from "@khanacademy/wonder-blocks-labeled-field";
+import * as React from "react";
+
+import ScrolllessNumberTextField from "../../../components/scrollless-number-text-field";
+import styles from "../image-editor.module.css";
+import {
+    wbFieldStyles,
+    getOtherSideLengthWithPreservedAspectRatio,
+} from "../utils";
+
+import type {Props as ImageEditorProps} from "../image-editor";
+import type {PerseusImageBackground} from "@khanacademy/perseus-core";
+
+interface Props {
+    backgroundImage: PerseusImageBackground;
+    onChange: ImageEditorProps["onChange"];
+}
+
+export default function ImageDimensionsInput({
+    backgroundImage,
+    onChange,
+}: Props) {
+    function handleWidthChange(newWidth: string) {
+        const newHeight = getOtherSideLengthWithPreservedAspectRatio(
+            backgroundImage.width!, // current side (width)
+            backgroundImage.height!, // other side (height)
+            Number(newWidth), // new side (width)
+        );
+
+        if (isNaN(newHeight)) {
+            return;
+        }
+
+        onChange({
+            backgroundImage: {
+                ...backgroundImage,
+                width: Number(newWidth),
+                height: newHeight,
+            },
+        });
+    }
+
+    function handleHeightChange(newHeight: string) {
+        const newWidth = getOtherSideLengthWithPreservedAspectRatio(
+            backgroundImage.height!, // current side (height)
+            backgroundImage.width!, // other side (width)
+            Number(newHeight), // new side (height)
+        );
+
+        if (isNaN(newWidth)) {
+            return;
+        }
+
+        onChange({
+            backgroundImage: {
+                ...backgroundImage,
+                height: Number(newHeight),
+                width: newWidth,
+            },
+        });
+    }
+
+    return (
+        <div className={styles.dimensionsContainer}>
+            <LabeledField
+                label="Width"
+                field={
+                    <ScrolllessNumberTextField
+                        value={backgroundImage.width?.toString() ?? ""}
+                        onChange={handleWidthChange}
+                    />
+                }
+                styles={wbFieldStyles}
+            />
+            <span className={styles.xSpan}>x</span>
+            <LabeledField
+                label="Height"
+                field={
+                    <ScrolllessNumberTextField
+                        value={backgroundImage.height?.toString() ?? ""}
+                        onChange={handleHeightChange}
+                    />
+                }
+                styles={wbFieldStyles}
+            />
+        </div>
+    );
+}